+++ changelog.md
@@ -2,7 +2,7 @@
 
 ### [Latest]
 
-- Add container image documentation and generic Slurm submission scripts with config-driven component enumeration [#160](https://github.com/umami-hep/umami-preprocessing/pull/160)
+- Add container image documentation and generic Slurm/HTCondor submission scripts with config-driven component enumeration [#160](https://github.com/umami-hep/umami-preprocessing/pull/160)
 ### [v0.3.2](https://github.com/umami-hep/umami-preprocessing/releases/tag/v0.3.2) (04.08.2026)
 
 - Add new configs for central dataset [#159](https://github.com/umami-hep/umami-preprocessing/pull/159)